move interpreter and minivm compiler to seperate files
